requirements.txt
appnope==0.1.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
coverage==4.5.4
decorator==4.4.1
entrypoints==0.3
flake8==3.7.9
google-api-core==1.14.3
google-auth==1.7.0
google-cloud-bigquery==1.21.0
google-cloud-core==1.0.3
google-cloud-dataproc==0.6.1
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
idna==2.8
importlib-metadata==0.23
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
joblib==0.14.0
jsonschema==3.1.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==7.2.0
-e git+git@github.com:mozilla/messaging-system-personalization-experiment-1-numbermuncher.git@5b44909f8b1392efce6490e2b42a2ac9be5e8477#egg=mozilla_cfr_personalization
numpy==1.17.3
packaging==19.2
pandas==0.25.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.0
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
py==1.8.0
py4j==0.10.7
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.4.4
pyparsing==2.4.4
pyrsistent==0.15.5
pyspark==2.4.4
pytest==5.2.2
pytest-cov==2.8.1
python-dateutil==2.8.1
python-decouple==3.1
pytz==2019.3
requests==2.22.0
requests-mock==1.7.0
rope==0.14.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.13.0
sklearn==0.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
